removed extra classes
